refactor(arn): one file per provider transformer
transformer.rs held the trait, the registry and four provider
translations in 727 lines. Adding Scaleway meant editing the same file
as AWS, GCP and Azure, and reading any one of them meant scrolling past
the other three.

Split so each provider owns a file, tests included: aws.rs, gcp.rs,
azure.rs, scaleway.rs. mod.rs keeps what is genuinely shared — the
ArnTransformer contract and get_transformer — and the one test that
covers the registry rather than a provider.

Moving the tests matters as much as moving the code. Leaving the AWS
tests in mod.rs would have contradicted the point: a provider you cannot
read, change and verify in one file is not isolated.

arn/mod.rs is untouched, so every re-export keeps its path and no
consumer sees this. 1440 tests pass, clippy clean.

What this does not address, and #103 raised the same thing from another
angle: wami-core still knows AWS, GCP, Azure and Scaleway by name.
Organising those four into separate files does not change that the core
carries provider knowledge at all — the crates/cloud-provider/ crates
exist and would be their natural home, but the registry here cannot
reach them without inverting the dependency.

Closes #66
